Extract on_access from __init__
