+++ cmk/gui/userdb/__init__.py
@@ -93,6 +93,7 @@ from ._check_credentials import (
 from ._check_credentials import user_exists as user_exists
 from ._check_credentials import user_exists_according_to_profile as user_exists_according_to_profile
 from ._check_credentials import user_locked as user_locked
+from ._on_access import on_access as on_access
 from ._two_factor import disable_two_factor_authentication as disable_two_factor_authentication
 from ._two_factor import is_two_factor_backup_code_valid as is_two_factor_backup_code_valid
 from ._two_factor import is_two_factor_login_enabled as is_two_factor_login_enabled
@@ -176,32 +177,6 @@ def _get_attributes(
     return selector(connection) if connection else []
 
 
-def _check_login_timeout(username: UserId, session_duration: float, idle_time: float) -> None:
-    _handle_max_duration(username, session_duration)
-    _handle_idle_timeout(username, idle_time)
-
-
-def _handle_max_duration(username: UserId, session_duration: float) -> None:
-    if (
-        max_duration := active_config.session_mgmt.get("max_duration", {}).get("enforce_reauth")
-    ) and session_duration > max_duration:
-        raise MKAuthException(
-            f"{username} login timed out (Maximum session duration of {max_duration / 60} minutes exceeded)"
-        )
-
-
-def _handle_idle_timeout(username: UserId, idle_time: float) -> None:
-    idle_timeout = load_custom_attr(
-        user_id=username, key="idle_timeout", parser=convert_idle_timeout
-    )
-    if idle_timeout is None:
-        idle_timeout = active_config.session_mgmt.get("user_idle_timeout")
-    if idle_timeout is not None and idle_timeout is not False and idle_time > idle_timeout:
-        raise MKAuthException(
-            f"{username} login timed out (Maximum inactivity of {idle_timeout / 60} minutes exceeded)"
-        )
-
-
 # userdb.need_to_change_pw returns either None or the reason description why the
 # password needs to be changed
 def need_to_change_pw(username: UserId, now: datetime) -> str | None:
@@ -276,25 +251,3 @@ def on_failed_login(username: UserId, now: datetime) -> None:
             log_msg_until_locked,
             request.remote_ip,
         )
-
-
-def on_access(username: UserId, session_id: str, now: datetime) -> None:
-    """
-
-    Raises:
-        - MKAuthException: when the session given by session_id is not valid
-        - MKAuthException: when the user has been idle for too long
-
-    """
-    session_infos = load_session_infos(username)
-    if not is_valid_user_session(username, session_infos, session_id):
-        raise MKAuthException("Invalid user session")
-
-    # Check whether there is an idle timeout configured, delete cookie and
-    # require the user to renew the log when the timeout exceeded.
-    session_info = session_infos[session_id]
-    _check_login_timeout(
-        username,
-        now.timestamp() - session_info.started_at,
-        now.timestamp() - session_info.last_activity,
-    )
